pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentwall"
version = "0.1.0"
description = "A dotfile-driven firewall that protects the OS from destructive LLM agent tool calls"
readme = "docs/README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    { name = "Wissam El-Labban" },
    { name = "prad" },
]
keywords = ["security", "llm", "agent", "firewall", "ai-safety"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "watchdog>=3.0",
    "psutil>=5.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]
sandbox = [
    "fusepy>=3.0",
]

[project.scripts]
agentfirewall = "agentfirewall.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/agentfirewall"]

[tool.pytest.ini_options]
testpaths = ["tests"]
